requirements.txt
annotated-types==0.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astroid==3.2.4
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
crypto==1.4.1
dill==0.3.8
idna==3.7
isort==5.13.2
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
mysql-connector-python==9.0.0
Naked==0.1.32
platformdirs==4.2.2
pyaes==1.6.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pylint==3.2.6
pyperclip==1.9.0
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
shellescape==3.8.1
tabulate==0.9.0
tomlkit==0.13.0
typing_extensions==4.12.2
urllib3==2.2.2
